+++ CHANGELOG.md
@@ -36,6 +36,7 @@ hi
     - A double-tall tropical flower.
   - Lithops
     - A flowerbed-like block, sporting multiple colors of Lithops.
+- Fixed the ordering of Trailier Tales's new Granite, Diorite, and Andesite building blocks in the Creative Inventory in 1.21.2+. ([#25](https://github.com/FrozenBlock/TrailierTales/issues/25))
 
 ### Assets
 - Fixed Trailier Tales' crop models floating one pixel above Farmland blocks.